requirements.txt
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
clean-fid==0.1.35
click==8.1.8
cmake==4.0.3
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==45.0.5
cycler==0.12.1
datasets==3.0.1
descartes==1.1.0
diffusers==0.27.2
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.14.0
fire==0.7.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.22.2
idna==3.10
image_utilities==0.0.3
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
jaxtyping==0.2.19
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
numpy==1.24.4
nuscenes-devkit==1.1.11
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
orjson==3.10.15
oss2==2.17.0
packaging==24.2
pandas==2.0.3
peft==0.10.0
pillow==10.4.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==17.0.0
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.1.4
pyquaternion==0.9.9
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.2
rich==13.4.2
ruff==0.11.11
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
sentry-sdk==2.29.1
setproctitle==1.3.6
Shapely==1.8.5.post1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.15
tokenizers==0.19.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.65.2
transformers==4.40.1
triton==2.1.0
typeguard==4.4.0
typer==0.16.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.33.0
wandb==0.19.11
websockets==12.0
xxhash==3.5.0
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
